+++ src/mem/slice/mod.rs
@@ -1,18 +1,17 @@
 // devela::mem::slice
 //
 //! Slices of memory.
-#![doc = crate::doc_!(extends: slice)]
-#![doc = crate::doc_!(modules: crate::mem; slice)]
-#![doc = crate::doc_!(newline)]
-//!
+// #![doc = crate::doc_!(extends: slice)]
+// #![doc = crate::doc_!(modules: crate::mem; slice)]
+// #![doc = crate::doc_!(newline)]
 //
 
 mod ext;
-mod wrapper;
-pub use {ext::*, wrapper::*};
+mod namespace;
+pub use {ext::*, namespace::*};
 
 pub(crate) mod all {
     #[doc(inline)]
     #[allow(unused_imports)]
-    pub use super::{ext::*, wrapper::*};
+    pub use super::{ext::*, namespace::*};
 }