+++ DOCS/CHANGELOG.md
@@ -49,6 +49,7 @@ The format is based on [Keep a Changelog], and this project adheres to
   - `Float` and `ExtFloatConst` consts: `LOW_MARGIN`, `MEDIUM_MARGIN`, `HIGH_MARGIN`.
   - `Graph` methods: `edge_exists_unchecked`, `edge_remove`.
   - `NonValue*`: `is_max`, `is_min`, `[checked|strict|saturating|wrapping]_[add|sub]`.
+  - `Slice`: `from_mut`, `from_ref`, `from_raw_parts`, `from_raw_parts_mut`.
 - macros:
   - `id_seq!`, `type_marker!`, `type_resource!`,
   - `const_bool!`, `capture_first`, `capture_last`, `capture_tail_tuple!`, `impl_non_value!`.